--- zero-cost-templating/src/bin/if_else.rs
@@ -1,43 +1,0 @@
-#![deny(warnings)]
-#![feature(coroutines)]
-
-extern crate alloc;
-
-use std::pin::pin;
-
-use futures::StreamExt;
-use zero_cost_templating_macros::template_stream;
-
-#[template_stream("if_else.html.hbs")]
-pub async fn test() {
-    let template = if_else_initial0();
-    let template = if 1 == 5 {
-        let template = template.if_else_template0();
-        template.if_else_template2()
-    } else {
-        let template = template.if_else_template1();
-        template.if_else_template3()
-    };
-    template.if_else_template4();
-}
-
-#[tokio::main]
-pub async fn main() {
-    if option_env!("ZERO_COST_TEMPLATING_NO_EXPAND").is_some() {
-        return;
-    }
-    let mut actual = String::new();
-    let stream = test();
-    let mut stream = pin!(stream);
-    while let Some(value) = stream.next().await {
-        actual.push_str(&value);
-    }
-    assert_eq!(
-        "<!DOCTYPE html>\n<html lang=\"en\">\n  <head>\n    <meta charset=\"utf-8\">\n    \
-    <title>title</title>\n    <link rel=\"stylesheet\" href=\"style.css\">\n    \
-    <script src=\"script.js\"></script>\n  </head>\n  <body>\n    <h1>thetitle</h1>\n    \
-    <input type=\"hidden\" value=\"thetoken\">\n    <ul>\n    \n    </ul>\n    \
-    <span>2023</span>\n  </body>\n</html>",
-        actual
-    );
-}
